% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arrayRtojs}
\alias{arrayRtojs}
\title{Convert R array to JavaScript object}
\usage{
arrayRtojs(name, value, type = "character", digits = 3, value_array = TRUE)
}
\arguments{
\item{name}{The name of the JavaScript object.}

\item{value}{The R array to be converted.}

\item{type}{The data type of the values in the R array. Default is "character".}

\item{digits}{The number of digits to round the numeric values. Default is 3.}

\item{value_array}{Logical value indicating whether the values should be converted into an array in JavaScript. Default is TRUE.}
}
\value{
A string representing the JavaScript object.
}
\description{
This function converts an R array into a JavaScript object. It supports different data types such as character, date, and numeric.
}
