% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{Chunk}
\alias{Chunk}
\title{Chunk functions}
\description{
From a netCDF file, generate two versions with the same data but with different chunk
configurations. In one case, favor the retrieval of temporal series of the main
variable in each pixel, and in the other, favor the retrieval of planes for each date.
}
\details{
\tabular{ll}{
Version: \tab 1.0.0\cr
License: \tab GPL version 3 or newer\cr
}
}
\author{
Borja Latorre-Garcés \url{https://www.eead.csic.es/home/staffinfo?Id=215}; Soil and Water, EEAD, CSIC \url{https://www.eead.csic.es}
Fergus Reig-Gracia \url{http://fergusreig.es}; Environmental Hydrology, Climate and Human Activity Interactions, Geoenvironmental Processes, IPE, CSIC \url{https://www.ipe.csic.es/hidrologia-ambiental}
Eduardo Moreno-Lamana \url{https://apuntes.eduardofilo.es}; Environmental Hydrology, Climate and Human Activity Interactions, Geoenvironmental Processes, IPE, CSIC \url{https://www.ipe.csic.es/hidrologia-ambiental}
Daniel Vilas-Perulán \url{https://www.eead.csic.es/home/staffinfo?Id=754}; Soil and Water, EEAD, CSIC \url{https://www.eead.csic.es}
Manuel Arretxea-Iriarte; Physics of climate and climate change, IGEO, CSIC \url{https://igeo.ucm-csic.es/}
}
