% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-web.R
\name{Config-web}
\alias{Config-web}
\title{Config-web functions}
\description{
write web configuration
}
\details{
\tabular{ll}{
Version: \tab 1.0.0\cr
License: \tab GPL version 3 or newer\cr
}
}
\author{
Borja Latorre-Garcés \url{https://www.eead.csic.es/home/staffinfo?Id=215}; Soil and Water, EEAD, CSIC \url{http://www.eead.csic.es}
Fergus Reig-Gracia \url{http://fergusreig.es}; Environmental Hydrology, Climate and Human Activity Interactions, Geoenvironmental Processes, IPE, CSIC \url{http://www.ipe.csic.es/hidrologia-ambiental}
}
