% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-web.R
\name{config_web}
\alias{config_web}
\title{Web configuration}
\usage{
config_web(
  file,
  folder,
  epsg,
  formatdates,
  varmin,
  varmax,
  varName,
  infoJs = NA,
  lon_name,
  lat_name,
  time_name = "time",
  portion_suffix
)
}
\arguments{
\item{file}{Path to the netCDF file.}

\item{folder}{Path to the output folder.}

\item{epsg}{EPSG code for spatial projection.}

\item{formatdates}{Date format for time variables.}

\item{varmin}{Minimum value for the main variable.}

\item{varmax}{Maximum value for the main variable.}

\item{varName}{Name of the netCDF file without the .nc extension.}

\item{infoJs}{List with additional information in JavaScript format.}

\item{lon_name}{Name of the longitude dimension.}

\item{lat_name}{Name of the latitude dimension.}

\item{time_name}{Name of the time dimension.}

\item{portion}{Portion of the netCDF file to process.}
}
\value{
List with the generated web configuration.
}
\description{
This function generates the web configuration for a specific netCDF file.
}
