% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_artifacts}
\alias{generate_artifacts}
\title{Generate artifacts from NetCDF file}
\usage{
generate_artifacts(
  nc_root,
  out_root,
  nc_filename,
  portion_suffix,
  var_id,
  epsg,
  info_js,
  write = FALSE
)
}
\arguments{
\item{nc_root}{The root directory of the NetCDF file.}

\item{out_root}{The root directory where the artifacts will be generated.}

\item{nc_filename}{The name of the NetCDF file.}

\item{portion_suffix}{The suffix to be added to the generated artifact files.}

\item{var_id}{The variable ID.}

\item{epsg}{The EPSG code.}

\item{info_js}{The information for the JavaScript configuration.}

\item{write}{A logical value indicating whether to write the artifacts to disk. Default is FALSE.}
}
\value{
The information for the JavaScript configuration.
}
\description{
This function generates artifacts from a NetCDF file, including chunked NetCDF files and corresponding binary files.
}
