% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeArrays}
\alias{mergeArrays}
\title{merge arrays using aggregation function}
\usage{
mergeArrays(arr1, arr2, positions, aggregation_func)
}
\arguments{
\item{arr1}{array 1}

\item{arr2}{array 2}

\item{positions}{positions}

\item{aggregation_func}{aggregation function to use (e.g., min, max, sum)}
}
\value{
merged array
}
\description{
merge arrays using aggregation function
}
