% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_coords}
\alias{read_coords}
\title{Read coordinates from a NetCDF file}
\usage{
read_coords(nc, epsg)
}
\arguments{
\item{nc}{The NetCDF file object.}

\item{epsg}{The EPSG code specifying the coordinate reference system (CRS).}
}
\value{
A data frame containing the longitude and latitude coordinates.
}
\description{
This function reads the longitude and latitude coordinates from a NetCDF file
and returns them as a data frame.
}
