% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_epsg}
\alias{read_epsg}
\title{Read EPSG code from NetCDF file}
\usage{
read_epsg(nc)
}
\arguments{
\item{nc}{The NetCDF file object.}
}
\value{
The EPSG code as a numeric value.
}
\description{
This function reads the EPSG code from a NetCDF file.
}
