% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_times}
\alias{read_times}
\title{read_times Function}
\usage{
read_times(nc, formatdates)
}
\arguments{
\item{nc}{The NetCDF file object.}

\item{formatdates}{Optional parameter to specify the format of the output dates.}
}
\value{
A vector of Date objects representing the time values.
}
\description{
This function reads the time values from a NetCDF file and converts them to Date objects.
}
