% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readMinMax}
\alias{readMinMax}
\title{readMinMax Function}
\usage{
readMinMax(nc)
}
\arguments{
\item{nc}{The NetCDF file object.}
}
\value{
A list containing the minimum and maximum values of the variable.
}
\description{
This function reads the minimum and maximum values of a variable from a NetCDF file.
}
