% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{returnXYNames}
\alias{returnXYNames}
\title{Return the names of the X and Y dimensions in a NetCDF object}
\usage{
returnXYNames(nc)
}
\arguments{
\item{nc}{A NetCDF object.}
}
\value{
A list containing the names and positions of the X and Y dimensions.
}
\description{
This function takes a NetCDF object as input and returns the names of the X and Y dimensions.
If the dimensions are named "Y" and "X", the function will return these names along with their positions.
If the dimensions are named "latitude" and "longitude", the function will return "latitude" as the Y dimension name,
"longitude" as the X dimension name, and their positions.
If the dimensions are named "lat" and "lon", the function will return "lat" as the Y dimension name,
"lon" as the X dimension name, and their positions.
}
