% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timePosition}
\alias{timePosition}
\title{timePosition}
\usage{
timePosition(nc)
}
\arguments{
\item{nc}{A netCDF object.}
}
\value{
The position of the "time" dimension.
}
\description{
This function returns the position of the "time" dimension in a netCDF object.
}
