% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-web.R
\name{writeJs}
\alias{writeJs}
\title{Write JavaScript file with configuration data}
\usage{
writeJs(
  folder,
  infoJs,
  varTitle,
  legendTitle = "Legend",
  offsetType = "Q",
  sizeType = "I",
  minify = TRUE
)
}
\arguments{
\item{folder}{The folder where the JavaScript file will be saved.}

\item{infoJs}{A list containing the configuration data, including times, variable minimum and maximum values, latitude and longitude ranges, and EPSG code.}

\item{varTitle}{The title of the variable.}

\item{legendTitle}{The title of the legend (optional, default is "Legend").}

\item{offsetType}{The type of offset (optional, default is "Q").}

\item{sizeType}{The type of size (optional, default is "I").}
}
\value{
The generated JavaScript code as a character string.
}
\description{
This function writes a JavaScript file with configuration data for VisorServiciosClimaticos.
}
