% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-web.R
\name{writeJson}
\alias{writeJson}
\title{Write JSON string with configuration data}
\usage{
writeJson(
  folder,
  infoJs,
  varTitle,
  legendTitle = "Legend",
  ncPortionSuffix = c("_pen", "_can"),
  offsetType = "Q",
  sizeType = "I",
  minify = TRUE
)
}
\arguments{
\item{infoJs}{A list containing the configuration data, including times, variable minimum and maximum values, latitude and longitude ranges, and EPSG code.}

\item{varTitle}{The title of the variable.}

\item{legendTitle}{The title of the legend (optional, default is "Legend").}

\item{ncPortionSuffix}{A character vector specifying the portion suffixes for the variable (optional, default is c("_pen", "_can")).}

\item{offsetType}{The type of offset (optional, default is "Q").}

\item{sizeType}{The type of size (optional, default is "I").}

\item{minify}{Whether the JSON string should be minified (optional, default is TRUE).}
}
\value{
The generated JSON string.
}
\description{
This function generates a JSON string with configuration data similar to the JavaScript variables in writeJs.
}
